Formated whole weightfree folder
